pyproject.toml
[project]
name = "desafio-stalse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.3",
    "fastapi>=0.124.4",
    "kaggle>=1.8.2",
    "kagglehub>=0.3.13",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]
